+++ src/components/layout/LayoutWrapper.tsx
@@ -1,5 +1,6 @@
 "use client";
 import { useWindowSize } from "@/hooks";
+import { useRequestSocket } from "@/hooks/socket/useRequestSocket";
 import { useAuth } from "@/providers/AuthProvider";
 import { useSocketStore } from "@/store/socket.store";
 import { useEffect } from "react";
@@ -12,6 +13,10 @@ export const LayoutWrapper = ({ children }: { children: React.ReactNode }) => {
   const { initSocket, resetSocket } = useSocketStore();
   const { accessToken } = useAuth();
 
+  // 소켓 핸들러  /// 아래 훅은 소켓 이벤트 핸들러입니다 지우지마세요
+  useRequestSocket();
+
+  // 소켓 초기화
   useEffect(() => {
     if (!accessToken) return;
     initSocket(accessToken);
